test: split pr finalizer governance contracts
